Refactor webpack handling
