+++ package.json
@@ -55,9 +55,9 @@
 		"errorstacks": "^1.3.0",
 		"expect": "^24.9.0",
 		"jasmine-core": "^3.5.0",
-		"karma": "^5.1.0",
+		"karma": "^5.1.1",
 		"karma-chrome-launcher": "^3.1.0",
-		"karma-coverage": "^2.0.2",
+		"karma-coverage": "^2.0.3",
 		"karma-firefox-launcher": "^1.3.0",
 		"karma-jasmine": "^3.3.1",
 		"karma-min-reporter": "^0.1.0",
